Treated SwiftLint warnings as errors and cleared all violations
